Added: Surface Level Cleanup for BC1
