+++ projects/dxt-lossless-transform-bc1/Cargo.toml
@@ -29,11 +29,11 @@ pprof = { version = "0.14", features = ["flamegraph", "criterion"] }
 
 # Benchmark Stuff
 [[bench]]
-name = "bc1_transform"
-path = "benches/bc1_transform/main.rs"
+name = "split_blocks"
+path = "benches/split_blocks/main.rs"
 harness = false
 
 [[bench]]
-name = "bc1_detransform"
-path = "benches/bc1_detransform/main.rs"
+name = "unsplit_blocks"
+path = "benches/unsplit_blocks/main.rs"
 harness = false
