+++ projects/dxt-lossless-transform-bc1/benches/split_blocks/portable64.rs
@@ -1,10 +1,10 @@
 use criterion::{black_box, BenchmarkId};
-use dxt_lossless_transform_bc1::bc1::transform::*;
+use dxt_lossless_transform_bc1::bc1::split_colours::*;
 use safe_allocator_api::RawAlloc;
 
-fn bench_avx2_gather(b: &mut criterion::Bencher, input: &RawAlloc, output: &mut RawAlloc) {
+fn bench_portable64_shift(b: &mut criterion::Bencher, input: &RawAlloc, output: &mut RawAlloc) {
     b.iter(|| unsafe {
-        gather(
+        shift(
             black_box(input.as_ptr()),
             black_box(output.as_mut_ptr()),
             black_box(input.len()),
@@ -12,34 +12,13 @@ fn bench_avx2_gather(b: &mut criterion::Bencher, input: &RawAlloc, output: &mut 
     });
 }
 
-#[cfg(target_arch = "x86_64")]
-fn bench_avx2_gather_unroll_4(b: &mut criterion::Bencher, input: &RawAlloc, output: &mut RawAlloc) {
-    b.iter(|| unsafe {
-        gather_unroll_4(
-            black_box(input.as_ptr()),
-            black_box(output.as_mut_ptr()),
-            black_box(input.len()),
-        )
-    });
-}
-
-fn bench_avx2_permute(b: &mut criterion::Bencher, input: &RawAlloc, output: &mut RawAlloc) {
-    b.iter(|| unsafe {
-        permute(
-            black_box(input.as_ptr()),
-            black_box(output.as_mut_ptr()),
-            black_box(input.len()),
-        )
-    });
-}
-
-fn bench_avx2_permute_unroll_2(
+fn bench_portable64_shift_with_count(
     b: &mut criterion::Bencher,
     input: &RawAlloc,
     output: &mut RawAlloc,
 ) {
     b.iter(|| unsafe {
-        permute_unroll_2(
+        shift_with_count(
             black_box(input.as_ptr()),
             black_box(output.as_mut_ptr()),
             black_box(input.len()),
@@ -47,14 +26,13 @@ fn bench_avx2_permute_unroll_2(
     });
 }
 
-#[cfg(target_arch = "x86_64")]
-fn bench_avx2_permute_unroll_4(
+fn bench_portable64_shift_unroll_2(
     b: &mut criterion::Bencher,
     input: &RawAlloc,
     output: &mut RawAlloc,
 ) {
     b.iter(|| unsafe {
-        permute_unroll_4(
+        shift_unroll_2(
             black_box(input.as_ptr()),
             black_box(output.as_mut_ptr()),
             black_box(input.len()),
@@ -62,23 +40,13 @@ fn bench_avx2_permute_unroll_4(
     });
 }
 
-fn bench_avx2_shuffle_permute(b: &mut criterion::Bencher, input: &RawAlloc, output: &mut RawAlloc) {
-    b.iter(|| unsafe {
-        shuffle_permute(
-            black_box(input.as_ptr()),
-            black_box(output.as_mut_ptr()),
-            black_box(input.len()),
-        )
-    });
-}
-
-fn bench_avx2_shuffle_permute_unroll_2(
+fn bench_portable64_shift_with_count_unroll_2(
     b: &mut criterion::Bencher,
     input: &RawAlloc,
     output: &mut RawAlloc,
 ) {
     b.iter(|| unsafe {
-        shuffle_permute_unroll_2(
+        shift_with_count_unroll_2(
             black_box(input.as_ptr()),
             black_box(output.as_mut_ptr()),
             black_box(input.len()),
@@ -86,14 +54,55 @@ fn bench_avx2_shuffle_permute_unroll_2(
     });
 }
 
-#[cfg(target_arch = "x86_64")]
-fn bench_avx2_shuffle_permute_unroll_4(
+fn bench_portable64_shift_unroll_4(
     b: &mut criterion::Bencher,
     input: &RawAlloc,
     output: &mut RawAlloc,
 ) {
     b.iter(|| unsafe {
-        shuffle_permute_unroll_4(
+        shift_unroll_4(
+            black_box(input.as_ptr()),
+            black_box(output.as_mut_ptr()),
+            black_box(input.len()),
+        )
+    });
+}
+
+fn bench_portable64_shift_with_count_unroll_4(
+    b: &mut criterion::Bencher,
+    input: &RawAlloc,
+    output: &mut RawAlloc,
+) {
+    b.iter(|| unsafe {
+        shift_with_count_unroll_4(
+            black_box(input.as_ptr()),
+            black_box(output.as_mut_ptr()),
+            black_box(input.len()),
+        )
+    });
+}
+
+fn bench_portable64_shift_unroll_8(
+    b: &mut criterion::Bencher,
+    input: &RawAlloc,
+    output: &mut RawAlloc,
+) {
+    b.iter(|| unsafe {
+        shift_unroll_8(
+            black_box(input.as_ptr()),
+            black_box(output.as_mut_ptr()),
+            black_box(input.len()),
+        )
+    });
+}
+
+fn bench_portable64_shift_with_count_unroll_8(
+    b: &mut criterion::Bencher,
+    input: &RawAlloc,
+    output: &mut RawAlloc,
+) {
+    b.iter(|| unsafe {
+        shift_with_count_unroll_8(
             black_box(input.as_ptr()),
             black_box(output.as_mut_ptr()),
             black_box(input.len()),
@@ -109,51 +118,52 @@ pub(crate) fn run_benchmarks(
     important_benches_only: bool,
 ) {
     group.bench_with_input(
-        BenchmarkId::new("avx2 shuffle_permute unroll 2", size),
+        BenchmarkId::new("portable64 shift no-unroll", size),
         &size,
-        |b, _| bench_avx2_shuffle_permute_unroll_2(b, input, output),
+        |b, _| bench_portable64_shift(b, input, output),
+    );
+
+    group.bench_with_input(
+        BenchmarkId::new("portable64 shift_with_count no-unroll", size),
+        &size,
+        |b, _| bench_portable64_shift_with_count(b, input, output),
+    );
+
+    group.bench_with_input(
+        BenchmarkId::new("portable64 shift unroll-8", size),
+        &size,
+        |b, _| bench_portable64_shift_unroll_8(b, input, output),
+    );
+
+    group.bench_with_input(
+        BenchmarkId::new("portable64 shift_with_count unroll-8", size),
+        &size,
+        |b, _| bench_portable64_shift_with_count_unroll_8(b, input, output),
     );
 
     if !important_benches_only {
-        group.bench_with_input(BenchmarkId::new("avx2 gather", size), &size, |b, _| {
-            bench_avx2_gather(b, input, output)
-        });
-
         group.bench_with_input(
-            BenchmarkId::new("avx2 shuffle_permute", size),
+            BenchmarkId::new("portable64 shift unroll-2", size),
             &size,
-            |b, _| bench_avx2_shuffle_permute(b, input, output),
-        );
-
-        group.bench_with_input(BenchmarkId::new("avx2 permute", size), &size, |b, _| {
-            bench_avx2_permute(b, input, output)
-        });
-
-        #[cfg(target_arch = "x86_64")]
-        group.bench_with_input(
-            BenchmarkId::new("avx2 shuffle_permute unroll 4", size),
-            &size,
-            |b, _| bench_avx2_shuffle_permute_unroll_4(b, input, output),
+            |b, _| bench_portable64_shift_unroll_2(b, input, output),
         );
 
         group.bench_with_input(
-            BenchmarkId::new("avx2 permute unroll 2", size),
+            BenchmarkId::new("portable64 shift_with_count unroll-2", size),
             &size,
-            |b, _| bench_avx2_permute_unroll_2(b, input, output),
+            |b, _| bench_portable64_shift_with_count_unroll_2(b, input, output),
         );
 
-        #[cfg(target_arch = "x86_64")]
         group.bench_with_input(
-            BenchmarkId::new("avx2 gather unroll 4", size),
+            BenchmarkId::new("portable64 shift unroll-4", size),
             &size,
-            |b, _| bench_avx2_gather_unroll_4(b, input, output),
+            |b, _| bench_portable64_shift_unroll_4(b, input, output),
         );
 
-        #[cfg(target_arch = "x86_64")]
         group.bench_with_input(
-            BenchmarkId::new("avx2 permute unroll 4", size),
+            BenchmarkId::new("portable64 shift_with_count unroll-4", size),
             &size,
-            |b, _| bench_avx2_permute_unroll_4(b, input, output),
+            |b, _| bench_portable64_shift_with_count_unroll_4(b, input, output),
         );
     }
 }
