+++ projects/dxt-lossless-transform-bc1/benches/split_blocks/sse2.rs
@@ -1,5 +1,5 @@
 use criterion::{black_box, BenchmarkId};
-use dxt_lossless_transform_bc1::bc1::transform::*;
+use dxt_lossless_transform_bc1::bc1::split_colours::*;
 use safe_allocator_api::RawAlloc;
 
 fn bench_punpckhqdq_unroll_2(b: &mut criterion::Bencher, input: &RawAlloc, output: &mut RawAlloc) {
