+++ projects/dxt-lossless-transform-bc1/src/bc1/split_colours/mod.rs
@@ -16,6 +16,69 @@ pub mod avx2;
 #[cfg(any(target_arch = "x86_64", target_arch = "x86"))]
 pub use avx2::*;
 
+#[cfg(any(target_arch = "x86_64", target_arch = "x86"))]
+#[inline(always)]
+unsafe fn split_blocks_x86(input_ptr: *const u8, output_ptr: *mut u8, len: usize) {
+    #[cfg(not(feature = "no-runtime-cpu-detection"))]
+    {
+        // Runtime feature detection
+        let avx2 = std::is_x86_feature_detected!("avx2");
+        let sse2 = std::is_x86_feature_detected!("sse2");
+
+        if avx2 && len % 128 == 0 {
+            shuffle_permute_unroll_2(input_ptr, output_ptr, len);
+            return;
+        }
+
+        if sse2 && len % 64 == 0 {
+            shufps_unroll_4(input_ptr, output_ptr, len);
+            return;
+        }
+    }
+
+    #[cfg(feature = "no-runtime-cpu-detection")]
+    {
+        #[cfg(target_feature = "avx2")]
+        if len % 128 == 0 {
+            shuffle_permute_unroll_2(input_ptr, output_ptr, len);
+            return;
+        }
+
+        #[cfg(target_feature = "sse2")]
+        if len % 64 == 0 {
+            shufps_unroll_4(input_ptr, output_ptr, len);
+            return;
+        }
+    }
+
+    // Fallback to portable implementation
+    u32(input_ptr, output_ptr, len)
+}
+
+/// Split BC1 blocks from standard interleaved format to separated color/index format
+/// using the best known implementation for the current CPU.
+///
+/// # Safety
+///
+/// - input_ptr must be valid for reads of len bytes
+/// - output_ptr must be valid for writes of len bytes
+/// - len must be divisible by 8
+/// - It is recommended that input_ptr and output_ptr are at least 16-byte aligned (recommended 32-byte align)
+#[inline]
+pub unsafe fn split_blocks(input_ptr: *const u8, output_ptr: *mut u8, len: usize) {
+    debug_assert!(len % 8 == 0);
+
+    #[cfg(any(target_arch = "x86_64", target_arch = "x86"))]
+    {
+        split_blocks_x86(input_ptr, output_ptr, len)
+    }
+
+    #[cfg(not(any(target_arch = "x86_64", target_arch = "x86")))]
+    {
+        split_colours::u32(input_ptr, output_ptr, len)
+    }
+}
+
 #[cfg(test)]
 pub mod tests {
     use super::*;
