+++ projects/dxt-lossless-transform-bc1/src/bc1/split_colours/portable64.rs
@@ -435,8 +435,8 @@ pub unsafe fn shift_with_count_unroll_8(input_ptr: *const u8, output_ptr: *mut u
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::bc1::transform::tests::generate_bc1_test_data;
-    use crate::bc1::transform::tests::transform_with_reference_implementation;
+    use crate::bc1::split_colours::tests::generate_bc1_test_data;
+    use crate::bc1::split_colours::tests::transform_with_reference_implementation;
     use rstest::rstest;
 
     // Define the function pointer type
