+++ projects/dxt-lossless-transform-bc1/src/bc1/split_colours/sse2.rs
@@ -438,8 +438,8 @@ pub unsafe fn shufps_unroll_8(input_ptr: *const u8, output_ptr: *mut u8, len: us
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::bc1::transform::tests::generate_bc1_test_data;
-    use crate::bc1::transform::tests::transform_with_reference_implementation;
+    use crate::bc1::split_colours::tests::generate_bc1_test_data;
+    use crate::bc1::split_colours::tests::transform_with_reference_implementation;
     use crate::testutils::allocate_align_64;
     use rstest::rstest;
