+++ projects/dxt-lossless-transform-bc1/src/bc1/unsplit_colours/sse2.rs
@@ -205,8 +205,8 @@ pub unsafe fn unpck_detransform_unroll_4(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::bc1::transform::tests::generate_bc1_test_data;
-    use crate::bc1::transform::u32;
+    use crate::bc1::split_colours::tests::generate_bc1_test_data;
+    use crate::bc1::split_colours::u32;
     use rstest::rstest;
 
     type DetransformFn = unsafe fn(*const u8, *mut u8, usize);
