+++ projects/dxt-lossless-transform-bc1/src/lib.rs
@@ -1,4 +1,4 @@
-#![doc = include_str!("../../../README.MD")]
+#![doc = include_str!(concat!("../", std::env!("CARGO_PKG_README")))]
 #![cfg_attr(not(feature = "std"), no_std)]
 
 pub mod bc1;
